api/requirements.txt
PyMuPDF==1.27.2.2
python-docx==1.2.0
